+++ experiments/long_runs/bucket.jl
@@ -19,6 +19,7 @@ using ClimaUtilities.ClimaArtifacts
 import Interpolations
 import ClimaUtilities.TimeVaryingInputs:
     TimeVaryingInput, LinearInterpolation, PeriodicCalendar
+import ClimaUtilities.TimeManager: ITime
 using ClimaDiagnostics
 using ClimaAnalysis
 import ClimaAnalysis.Visualize as viz
@@ -52,8 +53,14 @@ diagnostics_outdir = joinpath(root_path, "global_diagnostics")
 outdir =
     ClimaUtilities.OutputPathGenerator.generate_output_path(diagnostics_outdir)
 
-function setup_prob(t0, tf, Δt; outdir = outdir, nelements = (101, 7))
-
+function setup_prob(
+    t0,
+    tf,
+    Δt,
+    start_date;
+    outdir = outdir,
+    nelements = (101, 7),
+)
     earth_param_set = LP.LandParameters(FT)
     radius = FT(6378.1e3)
     depth = FT(3.5)
@@ -67,7 +74,6 @@ function setup_prob(t0, tf, Δt; outdir = outdir, nelements = (101, 7))
     surface_space = domain.space.surface
     subsurface_space = domain.space.subsurface
 
-    start_date = DateTime(2008)
     # Forcing data
     era5_artifact_path =
         ClimaLand.Artifacts.era5_land_forcing_data2008_folder_path(; context)
@@ -88,7 +94,7 @@ function setup_prob(t0, tf, Δt; outdir = outdir, nelements = (101, 7))
     z_0b = FT(1e-3)
     κ_soil = FT(1.5)
     ρc_soil = FT(2e6)
-    τc = FT(Δt)
+    τc = FT(float(Δt))
     α_snow = FT(0.8)
     albedo = PrescribedBaregroundAlbedo{FT}(α_snow, surface_space)
     bucket_parameters = BucketModelParameters(FT; albedo, z_0m, z_0b, τc)
@@ -119,7 +125,7 @@ function setup_prob(t0, tf, Δt; outdir = outdir, nelements = (101, 7))
         p,
     )
 
-    updateat = Array(t0:(3600 * 3):tf)
+    updateat = [promote(t0:(ITime(3600 * 3)):tf...)...]
     drivers = ClimaLand.get_drivers(bucket)
     updatefunc = ClimaLand.make_update_drivers(drivers)
 
@@ -164,6 +170,7 @@ function setup_and_solve_problem(; greet = false)
     years = 366days # years in seconds - 366 to make sure we capture at least full years
     tf = 1years # 2 years in seconds
     Δt = 900.0
+    start_date = DateTime(2008)
     nelements = (101, 7)
     if greet
         @info "Run: Global Bucket Model"
@@ -172,7 +179,11 @@ function setup_and_solve_problem(; greet = false)
         @info "Duration: $(tf - t0) s"
     end
 
-    prob, cb = setup_prob(t0, tf, Δt; nelements)
+    t0 = ITime(t0, epoch = start_date)
+    tf = ITime(tf, epoch = start_date)
+    Δt = ITime(Δt, epoch = start_date)
+    t0, tf, Δt = promote(t0, tf, Δt)
+    prob, cb = setup_prob(t0, tf, Δt, start_date; nelements)
 
     # Define timestepper and ODE algorithm
     timestepper = CTS.RK4()